feat: product filtering components reorganized or moved to separate modules directories and update imports
