+++ src/modules/productFilters/ui/PriceFilter.tsx
